requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
sniffio==1.3.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
yarl==1.8.2
